마이페이지(사진관리, 찜관리, 장바구니) 로그인 체크

+++ src/com/dahami/newsbank/web/servlet/MypageCart.java
@@ -8,11 +8,13 @@ import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.dahami.newsbank.dto.PhotoDTO;
 import com.dahami.newsbank.web.dao.BookmarkDAO;
 import com.dahami.newsbank.web.dao.CartDAO;
 import com.dahami.newsbank.web.dto.CartDTO;
+import com.dahami.newsbank.web.dto.MemberDTO;
 
 /**
  * Servlet implementation class MypageCart
@@ -41,23 +43,41 @@ public class MypageCart extends NewsbankServletBase {
 		request.setCharacterEncoding("UTF-8");
 		response.getWriter().append("Served at: ").append(request.getContextPath());
 		
-		String action = request.getParameter("action") == null ? "" : request.getParameter("action");
-		String member_seq = "1002";		
-		String uciCode = request.getParameter("uciCode");
-		CartDAO cartDAO = new CartDAO();
-		List<CartDTO> cartList = cartDAO.cartList(member_seq);
-		request.setAttribute("cartList", cartList);
-		
-		if(action.equals("delete")) {
-			cartDAO.deleteCart(member_seq, uciCode);
-		}else if(action.equals("bookmark")) {
-			String bookName = "기본그룹"; // 기본값
-			BookmarkDAO bookmarkDAO = new BookmarkDAO();
-			bookmarkDAO.insert(member_seq, uciCode, bookName);
+		HttpSession session = request.getSession();
+		MemberDTO MemberInfo = (MemberDTO) session.getAttribute("MemberInfo");
+
+		if (MemberInfo != null) {
+			
+			boolean mypageAuth = false;
+			if (session.getAttribute("mypageAuth") != null) {
+				mypageAuth = (boolean) session.getAttribute("mypageAuth");
+			}
+			if (mypageAuth == false) {
+				// 이전에 my page 비밀번호 입력했는지 체크
+				response.sendRedirect("/auth.mypage");
+			} else {
+				
+				String action = request.getParameter("action") == null ? "" : request.getParameter("action");
+				String member_seq = String.valueOf(MemberInfo.getSeq());
+				String uciCode = request.getParameter("uciCode");
+				CartDAO cartDAO = new CartDAO();
+				List<CartDTO> cartList = cartDAO.cartList(member_seq);
+				request.setAttribute("cartList", cartList);
+				
+				if(action.equals("delete")) {
+					cartDAO.deleteCart(member_seq, uciCode);
+				}else if(action.equals("bookmark")) {
+					String bookName = "기본그룹"; // 기본값
+					BookmarkDAO bookmarkDAO = new BookmarkDAO();
+					bookmarkDAO.insert(member_seq, uciCode, bookName);
+				}
+				
+				RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/mypage_cart.jsp");
+				dispatcher.forward(request, response);
+			}
+		}else {
+			response.sendRedirect("/login");
 		}
-		
-		RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/mypage_cart.jsp");
-		dispatcher.forward(request, response);
 	}
 
 	/**
